:sparkles: Update factories to generate more data. Fix typo in file name.
