+++ src/objects/tests/test_filters.py
@@ -5,7 +5,7 @@ from rest_framework.test import APITestCase
 
 from objects.accounts.constants import PermissionModes
 from objects.accounts.tests.factories import ObjectPermissionFactory
-from objects.core.tests.factores import ObjectFactory, ObjectRecordFactory
+from objects.core.tests.factories import ObjectFactory, ObjectRecordFactory
 from objects.utils.test import TokenAuthMixin
 
 OBJECT_TYPE = "https://example.com/objecttypes/v1/types/abc109"
