+++ src/objects/tests/test_geo_headers.py
@@ -5,7 +5,7 @@ from rest_framework.test import APITestCase
 
 from objects.accounts.constants import PermissionModes
 from objects.accounts.tests.factories import ObjectPermissionFactory
-from objects.core.tests.factores import ObjectFactory
+from objects.core.tests.factories import ObjectFactory
 from objects.utils.test import TokenAuthMixin
 
 from .constants import GEO_READ_KWARGS, POLYGON_AMSTERDAM_CENTRUM
